restructured the image storage class and media data managers
